Update patch set 4
Patch Set 4:

For example, if I try something like this:

[Filter "*"]
  reviewer = UserA
[Filter "branch:master"]
  reviewer = UserB

and then push a change, then only UserA gets tied to the change.

And if I switch the filters like so, 

[Filter "branch:master"]
  reviewer = UserB
[Filter "*"]
  reviewer = UserA

then only UserB gets tied to the change.

Also from reading the code, it seems like you intended to have
[Filter ""], but this filter doesn't work

Patch-set: 4