RFC: add suggestions when user types in reviewer box
When a user types in the reviewer box in the reviewer screen, suggestions
for users to add will show up. The suggestions include users who have
visibility of the project and will filter based on the users' full name.

Add a gerrit config option to set the max number of suggestions, defaults
to 10.

Change-Id: I0a8a5897fe08935631bb70d7f19b67b99b26d45c