Update patch set 6
Patch Set 6:

> Patch Set 4: Verified+1

I tested and suggestion works but the plugin is not adding the reviewers because it's not able to resolve the reviewer when it's a user, it works for groups.

I type 'sim' and it suggests the following user: 'user/simpleuser (simpleuser)'. The problem is that the AccountResolver is not able to resolve such a user.

Should the suggestion for users be the email instead?

Patch-set: 6
Label: Verified=0