Update patch set 8
Patch Set 8:

> > > I think we would fix the resolution of non internal ones?
 > >
 > > I am looking into that option.
 > 
 > Thanks. I cannot test and reproduce is though.

Actually, reproducing this problem is simple. Singleuser groups is a group backend that provides one group per user. When you install this plugin and type user/username, it will suggest the group but at resolution time, it will complain about not able to resolve such a group (Same error as for ldap groups).

I think we need to reuse Janice work on group reference from a plugin[1] but if I recall, this feature is limited to the plugin configuration in gerrit.config.

[1]https://gerrit-review.googlesource.com/#/c/70457

Patch-set: 8