Update patch set 8
Patch Set 8:

> > > Just to make sure I understand the problem correctly: By
 > > resolution
 > > > time you mean when the reviewer should be added?
 > >
 > > yes
 > 
 > Why is this the problem of this change? What happens, when you
 > enter "ldap/contrib-adm" without this change in config file?

Same error. Without this change, I would not have thought of using "ldap/contrib-adm" because I would have followed the documentation which does not mention external groups. The problem is that with this change, it suggest external groups which are not supported.

Supporting external groups should not necessarily be done in this change but in this change, we should limit the suggestions to the internals groups.

Patch-set: 8