Update patch set 8
Patch Set 8:

I would like to get this change merged.

David O., do you think it would be acceptable to filter out on the client side the groups that are not internal groups?

I looked into supporting external groups, not a small change since we need to change the way we write the config using group reference.

Current Group endpoint doesn't have any option to only suggest internal groups.

So the only quick solution I see to get this change in is to filter out groups that are not internal one.

Patch-set: 8