ChangeEventListener: Improve exception handling
If the configuration includes a username or email address that is
misspelled or has the incorrect case, the error log is filled up
with logs like:

Cannot resolve reviewer: User.Name
com.google.gerrit.extensions.restapi.UnprocessableEntityException: Group Not Found: User.Name
        at com.google.gerrit.server.group.GroupsCollection.parse(GroupsCollection.java:113)
        ...

This is misleading, because the problem is not actually caused by a
missing group.

Refactor the checks for account and group to handle exceptions
separately, and log appropriate warnings.

Change-Id: Ie29810e5edb9a5ebe9c603f288c17099aea875e9