Allow to disable REST API and UI
Add 2 new configuration options enableREST and enableUI which can be
set in the reviewers.config file in the site's etc folder.

When enableREST is disabled, the REST API is not exposed.

When enableUI is disabled, the 'Reviewers' menu is not added to the
project screen. This option is forcibly disabled when enableREST is
disabled.

Disabling the REST API and UI means that a project's reviewers.config
file may only be modified by directly pushing changes for review on the
project's refs/meta/config branch. This can be used to enforce code
reviews of the reviewer configuration.

Change-Id: I79ed94347a427385820c85c12f747432a2b4690b