Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > http://paste.openstack.org/show/657898/
> 
> Can you try:
> 
>   $ bazel build reviewers -s
> 
> and check the classpath of GWT execution and confirm that the classes from the error message are indeed missing in:
> 
>  external/gerrit_plugin_gwtui/jar/gerrit-plugin-gwtui-2.16-SNAPSHOT.jar
> 
> They should be there, I think.

The classes are in the jar file in the local m2 repository.  I can't check the file that it's complaining about because it's in a tmp location that ceases to exist after the build has stopped.

I have to go an do some other stuff now; I'll come back to this later.  Would appreciate if you could also try it and see if it works for you.

Patch-set: 1
Reviewer: David Pursehouse <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>