RFC: Implement reviewer suggestion
Since Gerrit version 2.14 it is possible for plugins to provide
reviewer suggestions [1].

Add a new plugin configuration reviewers.suggestOnly which, when
enabled, will make the plugin only suggest the reviewers that would
be added for the given change, instead of actually adding them.

Reuse the ChangeEventListener class which already implements all the
code necessary to get the reviewers from the config file, and rename
the class to simply Reviewers since it is no longer only an event
handler.

TODO: Documentation

[1] http://gerrit-documentation.storage.googleapis.com/Documentation/2.14/dev-plugins.html#reviewer-suggestion

Change-Id: Ibd153ddbe3d0f361edcb5488e258372baedec2d5