Add basic caching of reviewers config per project
The reviewers configuration is cached per project. The project
configuration is evicted from the cache when it is updated via the
REST API.

If the reviewers configuration is directly updated by git push to the
refs/meta/config branch, the cache is not evicted.

A further limitation is that project configuration can be inherited
from parent projects, but child project configurations are not
evited from the cache when a parent config is updated.

Both of these limitations will be addressed in follow-up commits.

Change-Id: I55bc64c75a701974d92f230adda702a206e90831