WiP/RFC: Support adding reviewers by blame
Adding reviewers according to the blame information is a useful feature,
but so far it is implemented in a separate plugin that is maintained
separately and has to be installed separately.

Add support for adding reviewers by blame, based on the functionality
in the reviewers-by-blame plugin.

In this initial version the by-blame functionality works alongside the
existing "by configuration" functionality. It is enabled by setting
the "reviewers.enableBlame" to true in the reviewers.config file in the
site's etc folder.

Suggestion of reviewers by blame is not implemented. If 'suggestOnly'
is enabled, suggestions are provided from the configured reviewers and
the blame functionality is effectively disabled.

TODO:
- Documentation
- Support for the maxReviewers and ignoreFileRegex configurations

Change-Id: I2bca805b09b2ce5d6e6ae63123b893904c0268bc