Support adding reviewers by blame
Adding reviewers according to the blame information is a useful feature,
but so far it is implemented in a separate plugin that is maintained
separately and has to be installed separately.

Add support for adding reviewers by blame, based on the functionality
in the reviewers-by-blame plugin.

In this initial version the by-blame functionality works alongside the
existing "by configuration" functionality. It is enabled by setting
the "reviewers.enableBlame" to true in the reviewers.config file in the
site's etc folder, and in the reviewers.config file on the project's
refs/meta/config. (It must be enabled both globally and per project).

Suggestion of reviewers by blame is not implemented. If 'suggestOnly'
is enabled, suggestions are provided from the configured reviewers and
the blame functionality is effectively disabled.

The number of reviewers added by blame can be limited by setting the
"reviewers.maxReviewers" value in the project's reviewers.config file.

Files can be excluded from the blame calculation by the "ignoreFiles"
setting which may be a regular expression.

When the reviewers are configured in the reviewers.config in addition
to enabling by-blame, those reviewers will always be added, and any
additional reviewers found by blame will also be added. If any of the
reviewers-by-blame overlap with the configured reviewers, they are not
excluded from the by-blame calculation, potentially resulting in fewer
reviewers being added. In a future change we may add steps to exclude
overlapping reviewers.

Reviewers by blame are weighted by the number of contributions found
in the blame results, and the "top reviewers" are used. In a future
change we may consider adding a threshold for this weighting.

TODO:
- Documentation

Change-Id: I2bca805b09b2ce5d6e6ae63123b893904c0268bc