Refactor ReviewersConfig to also handle the global config
Move the existing functionality of ReviewersConfig into a new inner
class, ForProject, which is now created by calling the forProject
method rather than invoking a Guice factory method.

Wrap all of the global configuration into ReviewersConfig, make it
a Singleton, and inject it at the places where it's needed rather
than using PluginConfigFactory every time.

A side effect of this change is that modifying the 'ignoreDrafts'
setting now only takes effect after reloading the plugin, whereas
before it was possible to change it without a reload. For all the
other global settings it makes no difference because they are only
referenced from the Module and therefore only effective at plugin
load/reload anyway.

Change-Id: I0af7d46b3e19f2623731315f90a177deb34d8307