reviewers: merge parent reviewers.config
ProjectLevelConfig.getWithInheritance gave us a reviewers.config where
reviewers from parent projects might be replaced by reviewers in child
projects.

This will introduce a custom method to lookup the parent configurations.
The reviewers in the parent projects are appended to the reviewers in
the child projects. This means that if you add yourself as a reviewer in
a parent project, there is no risk that you will be removed from the
reviewers list by a child projectes reviewers.config.

Bug: Issue 8468
Change-Id: I6af842f91b6faeb359808eaa72a6aeda2527985d