Update to new AccountResolver API
The new API provides a richer way to inspect the results of individual
account lookups. This change intends to keep the current behavior of
this plugin as similar as possible, although it's possible that in the
future we could improve the ReviewersResolver behavior by using the new
AccountResolver API differently.

Change-Id: I68b0c40bfc6eee88196bef7f625e35014521a121