Support complex filter queries
The check to see if the current change matches a filter query is
performed by querying the index with:

    change: $CHANGE_ID $FILTER_QUERY

And checking if the result is non-empty.

This works for simple queries like

    file:README.md

Since the built query would be:

    change:123 file:README.md

Which is equivalent to

    change:123 AND file:README.md

When the filter query is complex like:

    file:^pylib/my-service/.* OR file:^my-service

The built query would be:

    change:123 file:^pylib/my-service/.* OR file:^my-service

Which matches all changes where change-id = 123 that touches a file
that matches "^pylib/my-service/.*" but it also matches any change
that matches "file:^my-service". The result is often that all users
that exist in reviewers.config are added to all uploaded changes.

Change the query template to:

   change:$CHANGE_ID AND ($FILTER_QUERY)

Change-Id: I1cd98fc64e82074482859ac4877f4adb61646733