Update patch set 4
Patch Set 4:

@Thomas
This is something I played around with during the hackathon. David O suggested I'd ping you in.
I wont have time to continue working on this in the near future.
It's quite a mess but if you feel that you want to continue on it you're free to hack away.
The parent change was just a hack to cut down on build time.

Patch-set: 4
CC: Gerrit User 1029953 <1029953@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>