Update patch set 1
Patch Set 1:

While the problem fixed here is unrelalted to stable-2.16 and stable-3.0 branches, because Polymer 2 will only be activated on master, and expected to be included in Gerrit 3.1 for the first time, we could move that change to stable-2.16, correct?

One advantage of doing it, would be to avoid merge conflicts, wenn the code is changed on stable branches.

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>