Update patch set 1
Patch Set 1:

Hesitant to verify this as it is not verified on Polymer 2 only on Polymer 1.
Found yet another issue with Polymer 2 [1] when verifying this together with [2].
It shouldn't be related to this change, possibly to [2] but it won't appear without it since Polymer won't work at all if [2] is not applied.



[1] https://bugs.chromium.org/p/gerrit/issues/detail?id=11584
[2] https://gerrit-review.googlesource.com/c/plugins/reviewers/+/238611

Patch-set: 1