Update patch set 1
Patch Set 1:

This fixes the initial problem to run the plugin with polymer 2 but has also revealed another serioua Bug with Polymer 2 (see my comment on the merge change). It's bettet to have a not working UI then a destructive one. 
I did not have the time to solve that Bug yesterday, will try to sort it out today.

Patch-set: 1