Update patch set 1
Patch Set 1:

(1 comment)

Hi David,

sorry for the "messy" changeset, it was accidentally uploaded to the public gerrit instead of my organizations' private one. If you consider the feature "add reviewers as CC" useful, I will add an option for this to the plugin and fix the issues on this change, otherwise you can abandon it.

Patch-set: 1