Remove unnecessary abstraction of AddReviewers
Abstract class AddReviewers had only one implementation,
AddReviewersByConfiguration. This implementation wasn't configuration
specific, and all this implemetation did was to provide a Factory for
and a Set<Account.Id> of reviewers to the abstract class.
Getting the Set of reviewers from an abstract method has no benefits
to having them provided in the Constructor.

Feature: Issue 12610
Change-Id: I61d24b622f948fa71f85d24b76c315ea5cf477e8