Update patch set 1
Patch Set 1: Code-Review-1

I'm fine with changing this again. If you check the history it was like this laready. This was change for a reason. Not all queries worked. I don't remember what specific problems were fixed by that switch to query the index.

As Alice suggested in her overview in the issue we should document the restriction and mention the fact that we donÄt query the index.

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1