Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I'm fine with changing this again. If you check the history it was like this laready. This was change for a reason. Not all queries worked. I don't remember what specific problems were fixed by that switch to query the index.
> 
> As Alice suggested in her overview in the issue we should document the restriction and mention the fact that we donÄt query the index.

Is this really something we want to do?
To me it doesn't feel like it will be an improvement.
It's slower and adding reviewers is not that critical I think we can gamble with the off-chance that the index is out-dated.
And what effects would an out-dated index mean? The absolutely most common case is:
1. change is uploaded
2. Reviewers are added from patchset-created event.
What are the odds that the index gets outdated in a couple of ms?

Patch-set: 1