Don't query index when checking if change matches filter
This change limits the set of supported operands in a reviewer filter
query to those that do not rely on the index.

Since predicates currently doesn't know if they are depending on the
index or not we will have to maintain a static list of supported
query operators. This list is also unfortunately duplicated between
code and Documentation.

The checks plugin faces similar challenges and have already come
up with a solution for them, much of the validation of supported
operators and there enumeration are copied from there:
https://gerrit.googlesource.com/plugins/checks/+/ \
  08d8aad35e27f07f50095037bd99b173d8d7588e/ \
  java/com/google/gerrit/plugins/checks/CheckerQuery.java

This results in additional duplication until predicates can say
if they are index-dependent or not.

Feature: Issue 12610
Change-Id: I84bb11cd82d5f1c7539a96da2ec640b7a858fae6