Allow reviewer filter configuration without merging
This adds the global plugin configuration "mergeFilters", allowing
the configuration of reviewer filters without merging, e.g:

Project B is a child of Project A.
Project A defines the following filter:
	[filter "*"]
		reviewer = GroupA
Project B defines the following filter:
	[filter "*"]
		reviewer = GroupB

With mergeFilters = true (which is the default value, keeping current
behaviour), both GroupA and GroupB would be added as reviewers. With
mergeFilters = false, only GroupB would be added as reviewers.

Change-Id: I45a531f8c3718fdb991428ae470390a3c5c860d6