Match change predicates without using ChangeIndex
In the reviewers logic, we want to know if a configured filter
applies to a change. If so, we will add reviewers or CCs.

The trigger for this logic is a change update. Hence, we only
want to know if the filter applies to that change. We can do this
by just matching predicates directly and spare the query to the
change index thereby reducing latency.

Change-Id: Icd35b54eb6b0dad6eb357e7af75ab92aec9e3200