Don't add reviewers to merged changes
If a change goes directly from WIP to MERGED due to the push of a
WIP change (e.g.)
1. Upload a change with %wip.
2. Push the commit, thus closing the change automatically.

Reviewers are added and put in Attention-Set, because the change goes
public as soon as it is merged. Now they have notifications to act on a
change that is merged.

To avoid this, simply don't automatically add reviewers to merged
changes.

Change-Id: I8eb03a0507624a4e09f837e59eb764e8e6548c69